organizing, adding rust docs, etc.
