+++ wf_field/src/field.rs
@@ -1,5 +1,4 @@
 use super::field_definition::FieldDefinition;
-use wf_common::common::extract_bits;
 
 #[derive(Clone, Debug)]
 pub struct Field {
@@ -86,37 +85,6 @@ impl Field {
             .bytes
             .encoding
             .convert_to_bit_length(self.byte_length());
-    }
-
-    /**
-     * Extracts and decodes a Whiteflag message field from the binary buffer
-     * @param field the message field to be extracted and decoded
-     * @param startBit the bit where the encoded field is located in the buffer
-     * @return String with the decoded field value
-     * @throws WfCoreException if field connot be decoded
-     */
-    pub fn extract_message_field(
-        &mut self,
-        message_buffer: &[u8],
-        message_buffer_bit_length: usize,
-        start_bit: usize,
-    ) -> String {
-        let bit_length = if self.bit_length() >= 1 {
-            self.bit_length()
-        } else {
-            let mut bit_length = message_buffer_bit_length - start_bit;
-            bit_length -= bit_length % &self.definition.bytes.encoding.bit_length;
-            bit_length
-        };
-
-        let field_buffer: Vec<u8> = extract_bits(
-            message_buffer,
-            message_buffer_bit_length,
-            start_bit,
-            bit_length,
-        );
-
-        self.decode(&field_buffer)
     }
 }
